requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic[email]==2.5.0
anyio>=3.7.1,<4.0.0
starlette==0.27.0
sqlalchemy==2.0.23
alembic==1.12.1
redis==5.0.1
requests==2.31.0
httpx==0.25.2
pandas==2.1.3
matplotlib==3.8.2
Pillow==10.1.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
aiofiles==23.2.1
openai==1.3.7
gTTS==2.5.0
websockets==12.0
aiohttp==3.9.1
pytest==7.4.3
pytest-asyncio==0.21.1
